stelloglossa/requirements.txt
astropy>=6.0
psrqpy>=1.3
numpy>=1.26
pandas>=2.2
scipy>=1.12
networkx>=3.2
gensim>=4.3
requests>=2.31
anthropic>=0.40
openai>=1.50
google-generativeai>=0.8
voyageai>=0.3
python-dotenv>=1.0
jieba>=0.42
pytest>=8.0
